requirements.txt
# 桌面宠物项目依赖
# 迁移到PySide6的现代Qt桌面应用

# 核心依赖
PySide6>=6.4.0  # Qt6 Python绑定，用于现代GUI界面

# 开发模式依赖（热重载功能）
watchdog>=2.1.0  # 文件监控，用于热重载功能

# AI功能依赖
openai>=1.0.0  # OpenAI API客户端

# 专注功能依赖
Pillow>=9.0.0  # 图像处理，用于截图功能

# 其他可选依赖
# pystray>=0.17.0  # 系统托盘图标支持

# 打包工具依赖
pyinstaller>=5.0.0  # 用于打包成exe文件